foodgram/requirements.txt
asgiref==3.5.2
Django==4.1.4
django-filter==21.1
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.3.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.7
sqlparse==0.4.3
reportlab